// interfaces/models/establecimientos.interface.ts
export interface establecimientosAtributos {
    id ?: number;
    giro_id : number;
    ciudad_id :number;
    nombre :string;
    direccion: string;
    latitud : string;
    longitud: string;
    telefono: string;
    correo_electronico: string;
    url: string;
    esta_activo: boolean;
  }